Change name of TEST_RUN to get_finviz_articles
